refactor(examples): extract shared HGVS spec harvesting into examples/common
Lift the `sources`, `prefix` and `classify` modules out of
`generate_spec_fixture.rs` into `examples/common/spec_harvest.rs` so a second
generator can reuse the same discovery, prefixing and classification logic
rather than duplicating it. `examples/common/` has no `main.rs`, so Cargo does
not build it as an example target; each generator pulls it in with
`#[path = "common/spec_harvest.rs"] mod`.

Also add provenance the harvester previously discarded:

- code-span extraction now returns byte offsets (via `into_offset_iter`), and
  `Candidate` carries the resolved 1-based `source_line`;
- a new `negatives` module harvests the spec's prose negatives and classifies
  each by RFC 2119 strength, so a MUST-level rule can never be conflated with a
  SHOULD-level recommendation.

`hgvs_spec_normalization.json` is unchanged: `generate_spec_fixture --check`
passes byte-identical across the move, and `source_line` is deliberately not
serialized into that fixture.
